+++ asynctor/__init__.py
@@ -1,7 +1,7 @@
 from .aio import bulk_gather, gather, map_group, run, run_async, start_tasks, wait_for
 from .client import AsyncRedis
 from .timing import timeit
-from .utils import AsyncClientGenerator, AsyncTestClient, AttrDict, cache_attr
+from .utils import AsyncClientGenerator, AsyncTestClient, AttrDict, Shell, cache_attr
 
 __version__ = "0.8.5"
 __all__ = (
@@ -10,6 +10,7 @@ __all__ = (
     "AsyncRedis",
     "AsyncTestClient",
     "AttrDict",
+    "Shell",
     "bulk_gather",
     "cache_attr",
     "gather",