+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `StrEnum` to `asynctor.compat`
 - Add `Self` to `asynctor.compat`
 - Add `load_toml` to `asynctor.compat`
+- feat: support `from asynctor import Shell`
 
 ### [0.8.5](../../releases/tag/v0.8.5) - 2025-08-07